@@ -1,15 +1,11 @@
 package gorqlite
 
-/*
-	this file has low level stuff:
-
-	rqliteApiGet()
-	rqliteApiPost()
-
-	There is some code duplication between those and they should
-	probably be combined into one function.
-
-*/
+// this file has low level stuff:
+//
+// rqliteApiGet()
+// rqliteApiPost()
+//
+// nothing public here.
 
 import (
 	"bytes"
@@ -17,102 +13,147 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
+	"net/url"
 	"strings"
 )
 
-/* *****************************************************************
+type ParameterizedStatement struct {
+	Query     string
+	Arguments []interface{}
+}
 
-   method: rqliteApiGet() - for api_STATUS
-
-	- lowest level interface - does not do any JSON unmarshaling
-	- handles retries
-	- handles timeouts
-
- * *****************************************************************/
-
-func (conn *Connection) rqliteApiGet(ctx context.Context, apiOp apiOperation) ([]byte, error) {
-	var responseBody []byte
-	trace("%s: rqliteApiGet() called", conn.ID)
-
-	// only api_STATUS now - maybe someday BACKUP
-	if apiOp != api_STATUS && apiOp != api_NODES {
-		return responseBody, errors.New("rqliteApiGet() called for invalid api operation")
+// method: rqliteApiCall() - internally handles api calls,
+// not supposed to be used by other files
+//
+//   - handles retries
+//   - handles timeouts
+func (conn *Connection) rqliteApiCall(ctx context.Context, apiOp apiOperation, method string, requestBody []byte) ([]byte, error) {
+	// Verify that we have at least a single peer to which we can make the request
+	peers := conn.cluster.PeerList()
+	if len(peers) < 1 {
+		return nil, errors.New("don't have any cluster info")
 	}
+	trace("%s: I have a peer list %d peers long", conn.ID, len(peers))
 
-	// just to be safe, check this
-	peersToTry := conn.cluster.makePeerList()
-	if len(peersToTry) < 1 {
-		return responseBody, errors.New("I don't have any cluster info")
-	}
-	trace("%s: I have a peer list %d peers long", conn.ID, len(peersToTry))
+	// Keep list of failed requests to each peer, return in case all peers fail to answer
+	var failureLog []string
 
-	// failure log is used so that if all peers fail, we can say something
-	// about why each failed
-	failureLog := make([]string, 0)
+	for i, peer := range peers {
+		trace("%s: attempting to contact peer %d", conn.ID, i)
+		surl := conn.assembleURL(apiOp, peer)
 
-PeerLoop:
-	for peerNum, peerToTry := range peersToTry {
-		trace("%s: attemping to contact peer %d", conn.ID, peerNum)
-		// docs say default GET policy is up to 10 follows automatically
-		url := conn.assembleURL(apiOp, peerToTry)
-		req, err := http.NewRequestWithContext(ctx, "GET", url, nil)
+		// Prepare request
+		req, err := http.NewRequestWithContext(ctx, method, surl, bytes.NewBuffer(requestBody))
 		if err != nil {
 			trace("%s: got error '%s' doing http.NewRequest", conn.ID, err.Error())
-			failureLog = append(failureLog, fmt.Sprintf("%s failed due to %s", url, err.Error()))
-			continue PeerLoop
+			failureLog = append(failureLog, fmt.Sprintf("%s failed due to %s", redactURL(surl), err.Error()))
+			continue
 		}
 		trace("%s: http.NewRequest() OK", conn.ID)
 		req.Header.Set("Content-Type", "application/json")
 
-		client := conn.apiClient(true)
-		response, err := client.Do(req)
+		// Execute request using shared client
+		// We will close the response body as soon as we can to allow
+		// the TCP connection to escape back into client's pool
+		c := conn.apiClient(method == "GET")
+		response, err := c.Do(req)
 		if err != nil {
 			trace("%s: got error '%s' doing client.Do", conn.ID, err.Error())
-			failureLog = append(failureLog, fmt.Sprintf("%s failed due to %s", url, err.Error()))
-			continue PeerLoop
+			failureLog = append(failureLog, fmt.Sprintf("%s failed due to %s", redactURL(surl), err.Error()))
+			continue
 		}
-		trace("%s: client.Do() OK", conn.ID)
-		responseBody, err := ioutil.ReadAll(response.Body)
-		_ = response.Body.Close()
+
+		// Read response body even if not a successful answer to return a descriptive error message
+		responseBody, err := io.ReadAll(response.Body)
 		if err != nil {
 			trace("%s: got error '%s' doing ioutil.ReadAll", conn.ID, err.Error())
-			failureLog = append(failureLog, fmt.Sprintf("%s failed due to %s", url, err.Error()))
-			continue PeerLoop
+			failureLog = append(failureLog, fmt.Sprintf("%s failed due to %s", redactURL(surl), err.Error()))
+			_ = response.Body.Close()
+			continue
 		}
 		trace("%s: ioutil.ReadAll() OK", conn.ID)
-		if response.Status != "200 OK" {
+
+		// Check that we've got a successful answer
+		if response.StatusCode != http.StatusOK {
 			trace("%s: got code %s", conn.ID, response.Status)
-			failureLog = append(failureLog, fmt.Sprintf("%s failed, got: %s", url, response.Status))
-			continue PeerLoop
+			failureLog = append(failureLog, fmt.Sprintf("%s failed, got: %s, message: %s", redactURL(surl), response.Status, string(responseBody)))
+			_ = response.Body.Close()
+			continue
 		}
-		// if we got here, we succeeded
-		trace("%s: api call OK, returning", conn.ID)
+		_ = response.Body.Close()
+		trace("%s: client.Do() OK", conn.ID)
+
 		return responseBody, nil
 	}
 
-	// if we got here, all peers failed.  Let's build a verbose error message
-	var stringBuffer bytes.Buffer
-	stringBuffer.WriteString("tried all peers unsuccessfully. here are the results:\n")
+	// All peers have failed to answer us, build a verbose error message
+	var builder strings.Builder
+	builder.WriteString("tried all peers unsuccessfully. here are the results:\n")
 	for n, v := range failureLog {
-		stringBuffer.WriteString(fmt.Sprintf("   peer #%d: %s\n", n, v))
+		builder.WriteString(fmt.Sprintf("   peer #%d: %s\n", n, v))
 	}
-	return responseBody, errors.New(stringBuffer.String())
+	return nil, errors.New(builder.String())
+}
+
+// redactURL redacts URL from the given parameter to be
+// safely read by the client
+func redactURL(surl string) string {
+	u, err := url.Parse(surl)
+	if err != nil {
+		return ""
+	}
+	return u.Redacted()
+}
+
+//	   method: rqliteApiGet() - for api_STATUS and api_NODES
+//
+//		- lowest level interface - does not do any JSON unmarshalling
+//		- handles retries
+//		- handles timeouts
+func (conn *Connection) rqliteApiGet(ctx context.Context, apiOp apiOperation) ([]byte, error) {
+	var responseBody []byte
+	trace("%s: rqliteApiGet() called", conn.ID)
+
+	// Allow only api_STATUS and api_NODES now - maybe someday BACKUP
+	if apiOp != api_STATUS && apiOp != api_NODES {
+		return responseBody, errors.New("rqliteApiGet() called for invalid api operation")
+	}
+	return conn.rqliteApiCall(ctx, apiOp, "GET", nil)
 }
 
 // Statement enables use of parameterized sql statement.
 // The constructor issues a warning if the number of parameters does not match
 // the count of ? in the query.
 // example:
-//   x := NewStatement(
-//      "INSERT INTO Foo (id, name) VALUES ( ?, ? )",
-//      1,
-//      "bob")
+//
+//	x := NewStatement(
+//	   "INSERT INTO Foo (id, name) VALUES ( ?, ? )",
+//	   1,
+//	   "bob")
 type Statement struct {
 	Sql        string
 	Parameters []interface{}
 	Warning    string
+}
+
+func (s *Statement) P() ParameterizedStatement {
+	return ParameterizedStatement{
+		Query:     s.Sql,
+		Arguments: s.Parameters,
+	}
+}
+
+func makeParameterizedStatements(stmts []*Statement) []ParameterizedStatement {
+	if len(stmts) == 0 {
+		return nil
+	}
+	var ret []ParameterizedStatement
+	for _, s := range stmts {
+		ret = append(ret, s.P())
+	}
+	return ret
 }
 
 func NewStatement(sql string, params ...interface{}) *Statement {
@@ -153,95 +194,34 @@ func (s *Statement) MarshalJSON() ([]byte, error) {
 	return json.Marshal(all)
 }
 
-/* *****************************************************************
-
-   method: rqliteApiPost() - for api_QUERY and api_WRITE
-
-	- lowest level interface - does not do any JSON unmarshaling
- 	- handles 301s, etc.
-	- handles retries
-	- handles timeouts
-
-	it is called with an apiOperation type because the URL it will use varies
-	depending on the API operation type (api_QUERY vs. api_WRITE)
-
- * *****************************************************************/
-func (conn *Connection) rqliteApiPost(ctx context.Context, apiOp apiOperation, jStatements []byte) ([]byte, error) {
+//	   method: rqliteApiPost() - for api_QUERY and api_WRITE
+//
+//		- lowest level interface - does not do any JSON unmarshalling
+//		- handles retries
+//		- handles timeouts
+func (conn *Connection) rqliteApiPost(ctx context.Context, apiOp apiOperation, sqlStatements []ParameterizedStatement) ([]byte, error) {
 	var responseBody []byte
 
-	switch apiOp {
-	case api_QUERY:
-		trace("%s: rqliteApiGet() post called for a QUERY of %d statements", conn.ID, len(jStatements))
-	case api_WRITE:
-		trace("%s: rqliteApiGet() post called for a QUERY of %d statements", conn.ID, len(jStatements))
-	default:
-		return responseBody, errors.New("weird! called for an invalid apiOperation in rqliteApiPost()")
+	// Allow only api_QUERY and api_WRITE
+	if apiOp != api_QUERY && apiOp != api_WRITE {
+		return responseBody, errors.New("rqliteApiPost() called for invalid api operation")
 	}
 
-	// just to be safe, check this
-	peersToTry := conn.cluster.makePeerList()
-	if len(peersToTry) < 1 {
-		return responseBody, errors.New("I don't have any cluster info")
+	trace("%s: rqliteApiPost() called for a QUERY of %d statements", conn.ID, len(sqlStatements))
+
+	formattedStatements := make([][]interface{}, 0, len(sqlStatements))
+
+	for _, statement := range sqlStatements {
+		formattedStatement := make([]interface{}, 0, len(statement.Arguments)+1)
+		formattedStatement = append(formattedStatement, statement.Query)
+		formattedStatement = append(formattedStatement, statement.Arguments...)
+		formattedStatements = append(formattedStatements, formattedStatement)
 	}
 
-	// failure log is used so that if all peers fail, we can say something
-	// about why each failed
-	failureLog := make([]string, 0)
-
-PeerLoop:
-	for peerNum, peer := range peersToTry {
-		trace("%s: trying peer #%d", conn.ID, peerNum)
-
-		// we're doing a post, and the RFCs say that if you get a 301, it's not
-		// automatically followed, so we have to do that ourselves
-
-		responseStatus := "Haven't Tried Yet"
-		var url string
-		for responseStatus == "Haven't Tried Yet" || responseStatus == "301 Moved Permanently" {
-			url = conn.assembleURL(apiOp, peer)
-			req, err := http.NewRequestWithContext(ctx, "POST", url, bytes.NewBuffer(jStatements))
-			if err != nil {
-				trace("%s: got error '%s' doing http.NewRequest", conn.ID, err.Error())
-				failureLog = append(failureLog, fmt.Sprintf("%s failed due to %s", url, err.Error()))
-				continue PeerLoop
-			}
-			req.Header.Set("Content-Type", "application/json")
-			client := conn.apiClient(false)
-			response, err := client.Do(req)
-			if err != nil {
-				trace("%s: got error '%s' doing client.Do", conn.ID, err.Error())
-				failureLog = append(failureLog, fmt.Sprintf("%s failed due to %s", url, err.Error()))
-				continue PeerLoop
-			}
-			responseBody, err = ioutil.ReadAll(response.Body)
-			_ = response.Body.Close()
-			if err != nil {
-				trace("%s: got error '%s' doing ioutil.ReadAll", conn.ID, err.Error())
-				failureLog = append(failureLog, fmt.Sprintf("%s failed due to %s", url, err.Error()))
-				continue PeerLoop
-			}
-			responseStatus = response.Status
-			if responseStatus == "301 Moved Permanently" {
-				v := response.Header["Location"]
-				failureLog = append(failureLog, fmt.Sprintf("%s redirected me to %s", url, v[0]))
-				url = v[0]
-				continue PeerLoop
-			} else if responseStatus == "200 OK" {
-				trace("%s: api call OK, returning", conn.ID)
-				return responseBody, nil
-			} else {
-				trace("%s: got error in responseStatus: %s", conn.ID, responseStatus)
-				failureLog = append(failureLog, fmt.Sprintf("%s failed, got: %s", url, response.Status))
-				continue PeerLoop
-			}
-		}
+	body, err := json.Marshal(formattedStatements)
+	if err != nil {
+		return nil, err
 	}
 
-	// if we got here, all peers failed.  Let's build a verbose error message
-	var stringBuffer bytes.Buffer
-	stringBuffer.WriteString("tried all peers unsuccessfully. here are the results:\n")
-	for n, v := range failureLog {
-		stringBuffer.WriteString(fmt.Sprintf("   peer #%d: %s\n", n, v))
-	}
-	return responseBody, errors.New(stringBuffer.String())
+	return conn.rqliteApiCall(ctx, apiOp, "POST", body)
 }
