@@ -1,53 +1,49 @@
-/*
-	gorqlite
-	A golang database/sql driver for rqlite, the distributed consistent sqlite.
-
-	Copyright (c)2016 andrew fabbro (andrew@fabbro.org)
-
-	See LICENSE.md for license. tl;dr: MIT. Conveniently, the same licese as rqlite.
-
-	Project home page: https://github.com/raindo308/gorqlite
-
-	Learn more about rqlite at: https://github.com/rqlite/rqlite
-*/
+// Package gorqlite provieds a database/sql-like driver for rqlite,
+// the distributed consistent sqlite.
+//
+// Copyright (c)2016 andrew fabbro (andrew@fabbro.org)
+//
+// See LICENSE.md for license. tl;dr: MIT. Conveniently, the same license as rqlite.
+//
+// Project home page: https://github.com/raindo308/gorqlite
+//
+// Learn more about rqlite at: https://github.com/rqlite/rqlite
 package gorqlite
 
-/*
-	this file contains package-level stuff:
-		consts
-		init()
-		Open, TraceOn(), TraceOff()
-*/
+// this file contains package-level stuff:
+//   consts
+//   init()
+//   Open, TraceOn(), TraceOff()
 
 import (
 	"context"
 	"crypto/rand"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"net/url"
 	"strings"
 )
 
-/* *****************************************************************
-
-   const
-
- * *****************************************************************/
-
 type consistencyLevel int
 
 const (
-	cl_NONE consistencyLevel = iota
-	cl_WEAK
-	cl_STRONG
+	// ConsistencyLevelNone provides no consistency to other nodes.
+	ConsistencyLevelNone consistencyLevel = iota
+	// ConsistencyLevelWeak provides a weak consistency that guarantees the
+	// queries are sent to the leader.
+	ConsistencyLevelWeak
+	// ConsistencyLevelStrong provides a strong consistency and guarantees
+	// that queries are sent and received by other nodes.
+	ConsistencyLevelStrong
 )
 
 // used in several places, actually
-var consistencyLevelNames map[consistencyLevel]string
-var consistencyLevels map[string]consistencyLevel
+var (
+	consistencyLevelNames map[consistencyLevel]string
+	consistencyLevels     map[string]consistencyLevel
+)
 
 type apiOperation int
 
@@ -58,25 +54,18 @@ const (
 	api_NODES
 )
 
-/* *****************************************************************
-
-   init()
-
- * *****************************************************************/
-
 func init() {
-	traceOut = ioutil.Discard
+	traceOut = io.Discard
 
 	consistencyLevelNames = make(map[consistencyLevel]string)
-	consistencyLevelNames[cl_NONE] = "none"
-	consistencyLevelNames[cl_WEAK] = "weak"
-	consistencyLevelNames[cl_STRONG] = "strong"
+	consistencyLevelNames[ConsistencyLevelNone] = "none"
+	consistencyLevelNames[ConsistencyLevelWeak] = "weak"
+	consistencyLevelNames[ConsistencyLevelStrong] = "strong"
 
 	consistencyLevels = make(map[string]consistencyLevel)
-	consistencyLevels["none"] = cl_NONE
-	consistencyLevels["weak"] = cl_WEAK
-	consistencyLevels["strong"] = cl_STRONG
-
+	consistencyLevels["none"] = ConsistencyLevelNone
+	consistencyLevels["weak"] = ConsistencyLevelWeak
+	consistencyLevels["strong"] = ConsistencyLevelStrong
 }
 
 // Open creates and returns a "connection" to rqlite.
@@ -86,7 +75,8 @@ func init() {
 // config information.
 //
 // The URL is a comma-separated list of URLs to individual "seed" nodes of the cluster:
-//  http://frank:mySecret@host1:4001?level=strong&timeout=2,http://host2:4444,http://host3
+//
+//	http://frank:mySecret@host1:4001?level=strong&timeout=2,http://host2:4444,http://host3
 //
 // The first node URL may have additional settings (username, password, level, timeout)
 // and should be of the following form:
@@ -94,21 +84,22 @@ func init() {
 //	http://[username:password@]localhost[:4001][?level=none|weak|strong&timeout=1]
 //
 // Defaults:
-//  port:     4001
-//  username: empty
-//  password: empty
-//  level:    weak
-//  timeout:  2 (seconds)
-func Open(connURL string, client ...*http.Client) (Connection, error) {
+//
+//	port:     4001
+//	username: empty
+//	password: empty
+//	level:    weak
+//	timeout:  2 (seconds)
+func Open(connURL string, client ...*http.Client) (*Connection, error) {
 	return OpenContext(context.Background(), connURL, client...)
 }
 
-func OpenContext(ctx context.Context, connURL string, client ...*http.Client) (Connection, error) {
+func OpenContext(ctx context.Context, connURL string, client ...*http.Client) (*Connection, error) {
 	var cl *http.Client
 	if len(client) > 0 {
 		cl = client[0]
 	}
-	conn := Connection{
+	conn := &Connection{
 		client: cl,
 	}
 
@@ -122,7 +113,6 @@ func OpenContext(ctx context.Context, connURL string, client ...*http.Client) (C
 	trace("%s: Open() called for url: %s", conn.ID, connURL)
 
 	// set defaults
-	conn.timeout = 10
 	conn.hasBeenClosed = false
 
 	urls := strings.Split(connURL, ",")
@@ -140,53 +130,47 @@ func OpenContext(ctx context.Context, connURL string, client ...*http.Client) (C
 			return conn, err
 		}
 		var other peer
-		h, p, err := net.SplitHostPort(pu.Host)
+		_, _, err = net.SplitHostPort(pu.Host)
 		if err != nil {
-			other.hostname = pu.Host
-			other.port = "4001"
+			other = peer(fmt.Sprintf("%s:%d", pu.Host, 4001))
 		} else {
-			other.hostname = h
-			other.port = p
+			other = peer(pu.Host)
 		}
 		conn.cluster.otherPeers = append(conn.cluster.otherPeers, other)
 	}
+	conn.cluster.peerList = append(conn.cluster.peerList, conn.cluster.otherPeers...)
 
-	// call updateClusterInfo() to populate the cluster
-	// also tests the user's default
+	if !conn.disableClusterDiscovery {
+		// call updateClusterInfo() to re-populate the cluster and discover peers
+		// also tests the user's default
+		if err := conn.updateClusterInfo(ctx); err != nil {
+			return conn, err
+		}
+	}
 
-	err = conn.updateClusterInfo(ctx)
-
-	// and the err from updateClusterInfo() will be our err as well
-	return conn, err
+	return conn, nil
 }
 
-/* *****************************************************************
-
-	func: trace()
-
-	adds a message to the trace output
-
-	not a public function.  we (inside) can add - outside they can
-	only see.
-
-	Call trace as:     Sprintf pattern , args...
-
-	This is done so that the more expensive Sprintf() stuff is
-	done only if truly needed.  When tracing is off, calls to
-	trace() just hit a bool check and return.  If tracing is on,
-	then the Sprintfing is done at a leisurely pace because, well,
-	we're tracing.
-
-	Premature optimization is the root of all evil, so this is
-	probably sinful behavior.
-
-	Don't put a \n in your Sprintf pattern becuase trace() adds one
-
- * *****************************************************************/
-
+// trace adds a message to the trace output
+//
+// not a public function.  we (inside) can add - outside they can
+// only see.
+//
+// Call trace as:     Sprintf pattern , args...
+//
+// This is done so that the more expensive Sprintf() stuff is
+// done only if truly needed.  When tracing is off, calls to
+// trace() just hit a bool check and return.  If tracing is on,
+// then the Sprintf-ing is done at a leisurely pace because, well,
+// we're tracing.
+//
+// Premature optimization is the root of all evil, so this is
+// probably sinful behavior.
+//
+// Don't put a \n in your Sprintf pattern becuase trace() adds one
 func trace(pattern string, args ...interface{}) {
 	// don't do the probably expensive Sprintf() if not needed
-	if wantsTrace == false {
+	if !wantsTrace {
 		return
 	}
 
@@ -200,28 +184,22 @@ func trace(pattern string, args ...interface{}) {
 	_, _ = traceOut.Write([]byte(msg))
 }
 
-/*
-	TraceOn()
-
-	Turns on tracing output to the io.Writer of your choice.
-
-	Trace output is very detailed and verbose, as you might expect.
-
-	Normally, you should run with tracing off, as it makes absolutely
-	no concession to performance and is intended for debugging/dev use.
-*/
+// TraceOn turns on tracing output to the io.Writer of your choice.
+//
+// Trace output is very detailed and verbose, as you might expect.
+//
+// Normally, you should run with tracing off, as it makes absolutely
+// no concession to performance and is intended for debugging/dev use.
 func TraceOn(w io.Writer) {
 	traceOut = w
 	wantsTrace = true
 }
 
-/*
-	TraceOff()
-
-	Turns off tracing output.  Once you call TraceOff(), no further
-	info is sent to the io.Writer, unless it is TraceOn'd again.
-*/
+// TraceOff turns off tracing output. Once you call TraceOff(), no further
+// info is sent to the io.Writer, unless it is TraceOn'd again.
 func TraceOff() {
 	wantsTrace = false
-	traceOut = ioutil.Discard
+	traceOut = io.Discard
 }
+
+var _ = TraceOff
