@@ -21,7 +21,7 @@ func TestConnectionStringParse(t *testing.T) {
 	}
 
 	var conn Connection
-	if conn, err = parseUrl("https://user1:pass1@host1:4000/db?level=strong&timeout=1"); err !=nil {
+	if conn, err = parseUrl("https://user1:pass1@host1:4000/db?level=strong&timeout=1"); err != nil {
 		t.Error(err)
 	}
 
@@ -29,12 +29,12 @@ func TestConnectionStringParse(t *testing.T) {
 	requireBool(t, true, conn.wantsTransactions) // will always be true until settable
 	requireString(t, "user1", conn.username)
 	requireString(t, "pass1", conn.password)
-	requireString(t, "host1", conn.cluster.leader.hostname)
+	requireString(t, "host1:4000", string(conn.cluster.leader))
 	requireString(t, "strong", consistencyLevelNames[conn.consistencyLevel])
-	requireString(t, "4000", conn.cluster.leader.port)
 	requireInt(t, 1, conn.timeout)
 
-	if conn, err = parseUrl("http://user1:pass1@host1.foobar.com:4000/db?level=weak"); err !=nil {
+	//goland:noinspection HttpUrlsUsage
+	if conn, err = parseUrl("http://user1:pass1@host1.foobar.com:4000/db?level=weak"); err != nil {
 		t.Error(err)
 	}
 
@@ -42,9 +42,8 @@ func TestConnectionStringParse(t *testing.T) {
 	requireBool(t, true, conn.wantsTransactions) // will always be true until settable
 	requireString(t, "user1", conn.username)
 	requireString(t, "pass1", conn.password)
-	requireString(t, "host1.foobar.com", conn.cluster.leader.hostname)
+	requireString(t, "host1.foobar.com:4000", string(conn.cluster.leader))
 	requireString(t, "weak", consistencyLevelNames[conn.consistencyLevel])
-	requireString(t, "4000", conn.cluster.leader.port)
 	requireInt(t, defaultTimeout, conn.timeout)
 }
 
@@ -64,4 +63,57 @@ func requireInt(t *testing.T, expected int, actual int) {
 	if expected != actual {
 		t.Errorf("expected '%v', got '%v'", expected, actual)
 	}
+}
+
+func TestSetConsistencyLevel(t *testing.T) {
+	conn, err := Open(testUrl())
+	if err != nil {
+		t.Fatalf("failed to open connection: %v", err.Error())
+	}
+
+	t.Run("Less than none", func(t *testing.T) {
+		err := conn.SetConsistencyLevel(-1)
+		if err == nil {
+			t.Errorf("expected error, got nil")
+		}
+	})
+
+	t.Run("Greater than strong", func(t *testing.T) {
+		err := conn.SetConsistencyLevel(100)
+		if err == nil {
+			t.Errorf("expected error, got nil")
+		}
+	})
+
+	t.Run("None", func(t *testing.T) {
+		err := conn.SetConsistencyLevel(ConsistencyLevelNone)
+		if err != nil {
+			t.Errorf("expected nil, got %v", err)
+		}
+
+		currentLevel, err := conn.ConsistencyLevel()
+		if err != nil {
+			t.Errorf("unexpected error: %s", err.Error())
+		}
+
+		if currentLevel != "none" {
+			t.Errorf("expected currentLevel to be 'none', instead got %s", currentLevel)
+		}
+	})
+
+	conn.Close()
+}
+
+func TestSetExecutionWithTransaction(t *testing.T) {
+	conn, err := Open(testUrl())
+	if err != nil {
+		t.Fatalf("failed to open connection: %v", err.Error())
+	}
+
+	err = conn.SetExecutionWithTransaction(true)
+	if err != nil {
+		t.Errorf("expected nil, got %v", err)
+	}
+
+	conn.Close()
 }
