@@ -1,11 +1,5 @@
 package gorqlite
 
-/*
-	this file has
-		Write()
-		WriteResult and its methods
-*/
-
 import (
 	"context"
 	"encoding/json"
@@ -55,56 +49,105 @@ import (
 
  * *****************************************************************/
 
-/*
-WriteOne() is a convenience method that wraps Write() into a single-statement
-method.
-*/
-
+// WriteOne wraps Write() into a single-statement
+// method.
+//
+// WriteOne uses context.Background() internally; to specify the context, use WriteOneContext.
 func (conn *Connection) WriteOne(sqlStatement string) (wr WriteResult, err error) {
-	if conn.hasBeenClosed {
-		wr.Err = errClosed
-		return wr, errClosed
-	}
 	wra, err := conn.Write([]string{sqlStatement})
 	return wra[0], err
 }
 
-/*
-Write is used to perform DDL/DML in the database.  ALTER, CREATE, DELETE, DROP, INSERT, UPDATE, etc. all go through Write().
+// WriteOneContext wraps WriteContext() into a single-statement
+func (conn *Connection) WriteOneContext(ctx context.Context, sqlStatement string) (wr WriteResult, err error) {
+	wra, err := conn.WriteContext(ctx, []string{sqlStatement})
+	return wra[0], err
+}
 
-Write() takes an array of SQL statements, and returns an equal-sized array of WriteResults, each corresponding to the SQL
-statement that produced it.
+// WriteOneParameterized wraps WriteParameterized() into a single-statement method.
+//
+// WriteOneParameterized uses context.Background() internally; to specify the context, use WriteOneParameterizedContext.
+func (conn *Connection) WriteOneParameterized(statement ParameterizedStatement) (wr WriteResult, err error) {
+	wra, err := conn.WriteParameterized([]ParameterizedStatement{statement})
+	return wra[0], err
+}
 
-All statements are executed as a single transaction.
-
-Write() returns an error if one is encountered during its operation.  If it's something like a call to the rqlite API,
-then it'll return that error.  If one statement out of several has an error, it will return a generic "there were %d
-statement errors" and you'll have to look at the individual statement's Err for more info.
-*/
-func (conn *Connection) Write(sqlStatements []string) (results []WriteResult, err error) {
-	jStatements, err := json.Marshal(sqlStatements)
-	if err != nil {
-		return nil, err
-	}
-	return conn.write(context.Background(), jStatements)
+// WriteOneParameterizedContext wraps WriteParameterizedContext into
+// a single-statement method.
+func (conn *Connection) WriteOneParameterizedContext(ctx context.Context, statement ParameterizedStatement) (wr WriteResult, err error) {
+	wra, err := conn.WriteParameterizedContext(ctx, []ParameterizedStatement{statement})
+	return wra[0], err
 }
 
 func (conn *Connection) WriteStmt(ctx context.Context, sqlStatements ...*Statement) (results []WriteResult, err error) {
-	jStatements, err := json.Marshal(sqlStatements)
-	if err != nil {
-		return nil, err
-	}
-	return conn.write(ctx, jStatements)
+	return conn.WriteParameterizedContext(ctx, makeParameterizedStatements(sqlStatements))
 }
 
-func (conn *Connection) write(ctx context.Context, sqlStatements []byte) (results []WriteResult, err error) {
+// Write is used to perform DDL/DML in the database synchronously without parameters.
+//
+// Write uses context.Background() internally; to specify the context, use WriteContext.
+// To use Write with parameterized queries, use WriteParameterized.
+func (conn *Connection) Write(sqlStatements []string) (results []WriteResult, err error) {
+	parameterizedStatements := make([]ParameterizedStatement, 0, len(sqlStatements))
+	for _, sqlStatement := range sqlStatements {
+		parameterizedStatements = append(parameterizedStatements, ParameterizedStatement{
+			Query: sqlStatement,
+		})
+	}
+
+	return conn.WriteParameterized(parameterizedStatements)
+}
+
+// WriteContext is used to perform DDL/DML in the database synchronously without parameters.
+//
+// To use WriteContext with parameterized queries, use WriteParameterizedContext.
+func (conn *Connection) WriteContext(ctx context.Context, sqlStatements []string) (results []WriteResult, err error) {
+	parameterizedStatements := make([]ParameterizedStatement, 0, len(sqlStatements))
+	for _, sqlStatement := range sqlStatements {
+		parameterizedStatements = append(parameterizedStatements, ParameterizedStatement{
+			Query: sqlStatement,
+		})
+	}
+
+	return conn.WriteParameterizedContext(ctx, parameterizedStatements)
+}
+
+// WriteParameterized is used to perform DDL/DML in the database synchronously.
+//
+// WriteParameterized takes an array of SQL statements, and returns an equal-sized array of WriteResults,
+// each corresponding to the SQL statement that produced it.
+//
+// All statements are executed as a single transaction.
+//
+// WriteParameterized returns an error if one is encountered during its operation.
+// If it's something like a call to the rqlite API, then it'll return that error.
+// If one statement out of several has an error, it will return a generic
+// "there were %d statement errors" and you'll have to look at the individual statement's Err for more info.
+//
+// WriteParameterized uses context.Background() internally; to specify the context, use WriteParameterizedContext.
+func (conn *Connection) WriteParameterized(sqlStatements []ParameterizedStatement) (results []WriteResult, err error) {
+	return conn.WriteParameterizedContext(context.Background(), sqlStatements)
+}
+
+// WriteParameterizedContext is used to perform DDL/DML in the database synchronously.
+//
+// WriteParameterizedContext takes an array of SQL statements, and returns an equal-sized array of WriteResults,
+// each corresponding to the SQL statement that produced it.
+//
+// All statements are executed as a single transaction.
+//
+// WriteParameterizedContext returns an error if one is encountered during its operation.
+// If it's something like a call to the rqlite API, then it'll return that error.
+// If one statement out of several has an error, it will return a generic
+// "there were %d statement errors" and you'll have to look at the individual statement's Err for more info.
+func (conn *Connection) WriteParameterizedContext(ctx context.Context, sqlStatements []ParameterizedStatement) (results []WriteResult, err error) {
 	results = make([]WriteResult, 0)
 
 	if conn.hasBeenClosed {
 		var errResult WriteResult
-		errResult.Err = errClosed
+		errResult.Err = ErrClosed
 		results = append(results, errResult)
-		return results, errClosed
+		return results, ErrClosed
 	}
 
 	trace("%s: Write() for %d statements", conn.ID, len(sqlStatements))
@@ -129,14 +172,12 @@ func (conn *Connection) write(ctx context.Context, sqlStatements []byte) (result
 		return results, err
 	}
 
-	/*
-		at this point, we have a "results" section and
-		a "time" section.  we can igore the latter.
-	*/
+	// at this point, we have a "results" section and
+	// a "time" section.  we can igore the latter.
 
 	resultsArray, ok := sections["results"].([]interface{})
 	if !ok {
-		err = errors.New("Result key is missing from response")
+		err = errors.New("result key is missing from response")
 		trace("%s: sections[\"results\"] ERROR: %s", conn.ID, err)
 		var errResult WriteResult
 		errResult.Err = err
@@ -183,23 +224,105 @@ func (conn *Connection) write(ctx context.Context, sqlStatements []byte) (result
 	trace("%s: finished parsing, returning %d results", conn.ID, len(results))
 
 	if numStatementErrors > 0 {
-		return results, errors.New(fmt.Sprintf("there were %d statement errors", numStatementErrors))
+		return results, fmt.Errorf("there were %d statement errors", numStatementErrors)
 	} else {
 		return results, nil
 	}
 }
 
-/* *****************************************************************
+// QueueOne is a convenience method that wraps Queue into a single-statement.
+//
+// QueueOne uses context.Background() internally; to specify the context, use QueueOneContext.
+func (conn *Connection) QueueOne(sqlStatement string) (seq int64, err error) {
+	sqlStatements := make([]string, 0)
+	sqlStatements = append(sqlStatements, sqlStatement)
+	return conn.Queue(sqlStatements)
+}
 
-   type: WriteResult
+// QueueOneContext is a convenience method that wraps QueueContext into a single-statement
+func (conn *Connection) QueueOneContext(ctx context.Context, sqlStatement string) (seq int64, err error) {
+	return conn.QueueContext(ctx, []string{sqlStatement})
+}
 
- * *****************************************************************/
+// QueueOneParameterized is a convenience method that wraps QueueParameterized into a single-statement method.
+//
+// QueueOneParameterized uses context.Background() internally; to specify the context, use QueueOneParameterizedContext.
+func (conn *Connection) QueueOneParameterized(statement ParameterizedStatement) (seq int64, err error) {
+	return conn.QueueParameterized([]ParameterizedStatement{statement})
+}
 
-/*
-A WriteResult holds the result of a single statement sent to Write().
+// QueueOneParameterizedContext is a convenience method that wraps QueueParameterizedContext() into a single-statement method.
+func (conn *Connection) QueueOneParameterizedContext(ctx context.Context, statement ParameterizedStatement) (seq int64, err error) {
+	return conn.QueueParameterizedContext(ctx, []ParameterizedStatement{statement})
+}
 
-Write() returns an array of WriteResult vars, while WriteOne() returns a single WriteResult.
-*/
+// Queue is used to perform asynchronous writes to the rqlite database as defined in the documentation:
+// https://github.com/rqlite/rqlite/blob/master/DOC/QUEUED_WRITES.md
+//
+// Queue uses context.Background() internally; to specify the context, use QueueContext.
+// To use Queue with parameterized queries, use QueueParameterized.
+func (conn *Connection) Queue(sqlStatements []string) (seq int64, err error) {
+	return conn.QueueContext(context.Background(), sqlStatements)
+}
+
+// QueueContext is used to perform asynchronous writes to the rqlite database as defined in the documentation:
+// https://github.com/rqlite/rqlite/blob/master/DOC/QUEUED_WRITES.md
+//
+// To use QueueContext with parameterized queries, use QueueParameterizedContext.
+func (conn *Connection) QueueContext(ctx context.Context, sqlStatements []string) (seq int64, err error) {
+	parameterizedStatements := make([]ParameterizedStatement, 0)
+	for _, sqlStatement := range sqlStatements {
+		parameterizedStatements = append(parameterizedStatements, ParameterizedStatement{Query: sqlStatement})
+	}
+
+	return conn.QueueParameterizedContext(ctx, parameterizedStatements)
+}
+
+// QueueParameterized is used to perform asynchronous writes with parameterized queries
+// to the rqlite database as defined in the documentation:
+// https://github.com/rqlite/rqlite/blob/master/DOC/QUEUED_WRITES.md
+//
+// QueueParameterized uses context.Background() internally; to specify the context, use QueueParameterizedContext.
+func (conn *Connection) QueueParameterized(sqlStatements []ParameterizedStatement) (seq int64, err error) {
+	return conn.QueueParameterizedContext(context.Background(), sqlStatements)
+}
+
+// QueueParameterizedContext is used to perform asynchronous writes with parameterized queries
+// to the rqlite database as defined in the documentation:
+// https://github.com/rqlite/rqlite/blob/master/DOC/QUEUED_WRITES.md
+func (conn *Connection) QueueParameterizedContext(ctx context.Context, sqlStatements []ParameterizedStatement) (seq int64, err error) {
+	if conn.hasBeenClosed {
+		return 0, ErrClosed
+	}
+
+	trace("%s: Write() for %d statements", conn.ID, len(sqlStatements))
+
+	// Set queuing mode just for this call.
+	conn.wantsQueueing = true
+	defer func() {
+		conn.wantsQueueing = false
+	}()
+
+	response, err := conn.rqliteApiPost(ctx, api_WRITE, sqlStatements)
+	if err != nil {
+		trace("%s: rqliteApiCall() ERROR: %s", conn.ID, err.Error())
+		return 0, err
+	}
+	trace("%s: rqliteApiCall() OK", conn.ID)
+
+	var sections map[string]interface{}
+	err = json.Unmarshal(response, &sections)
+	if err != nil {
+		trace("%s: json.Unmarshal() ERROR: %s", conn.ID, err.Error())
+		return 0, err
+	}
+
+	return int64(sections["sequence_number"].(float64)), nil
+}
+
+// WriteResult holds the result of a single statement sent to Write().
+//
+// Write() returns an array of WriteResult vars, while WriteOne() returns a single WriteResult.
 type WriteResult struct {
 	Err          error // don't trust the rest if this isn't nil
 	Timing       float64
